Code works much better after changing it to int

+++ wavMONO/wavCPP/wavCPP.cpp
@@ -11,7 +11,7 @@ Version: 1.0
 
 #include "pch.h"
 
-void wavMONO(short* leftChannel, short* rightChannel, int startIndex, int endIndex) {
+void wavMONO(int* leftChannel, int* rightChannel, int startIndex, int endIndex) {
 
     for (int i = startIndex; i < endIndex; ++i) {
         leftChannel[i] = (leftChannel[i] + rightChannel[i]) / 2;
@@ -19,6 +19,6 @@ void wavMONO(short* leftChannel, short* rightChannel, int startIndex, int endInd
 
 }
 
-extern "C" __declspec(dllexport) void stereoToMono(short* leftChannel, short* rightChannel, int startIndex, int endIndex) {
+extern "C" __declspec(dllexport) void stereoToMono(int* leftChannel, int* rightChannel, int startIndex, int endIndex) {
 	wavMONO(leftChannel,rightChannel, startIndex, endIndex);
 }
